Extract `DepKindVTable` constructors to their own module
